permissions

+++ src/controllers/auth.js
@@ -51,8 +51,9 @@ module.exports = {
 
                     const accessInfo = {
                         key: process.env.ACCESS_KEY,
-                        time: '30s',
+                        time: process.env?.ACCESS_EXP || '30m',
                         data: {
+                            _id: user._id,
                             id: user.id,
                             username: user.username,
                             email: user.email,
@@ -64,7 +65,7 @@ module.exports = {
 
                     const refreshInfo = {
                         key: process.env.REFRESH_KEY,
-                        time: '3d',
+                        time: process.env?.REFRESH_EXP || '30m',
                         data: {
                             id: user.id,
                             password: user.password // encrypted password
@@ -129,7 +130,7 @@ module.exports = {
                     res.status(200).send({
                         error: false,
                         bearer: {
-                            access: jwt.sign(user.toJSON(), process.env.ACCESS_KEY, { expiresIn: '30m' })
+                            access: jwt.sign(user.toJSON(), process.env.ACCESS_KEY, { expiresIn: (process.env?.ACCESS_EXP || '30m') })
                         }
                     })
 
@@ -160,12 +161,23 @@ module.exports = {
 
         const auth = req.headers?.authorization // Token ...tokenKey...
         const tokenKey = auth ? auth.split(' ') : null // ['Token', '...tokenKey...']
-        const result = await Token.deleteOne({ token: tokenKey[1] })
 
-        res.send({
-            error: false,
-            message: 'Token deleted. Logout was OK.',
-            result
-        })
+
+        if (tokenKey[0] == 'Token') {
+
+            const result = await Token.deleteOne({ token: tokenKey[1] })
+            res.send({
+                error: false,
+                message: 'Token deleted. Logout was OK.',
+                result
+            })
+
+        } else {
+
+            res.send({
+                error: false,
+                message: 'JWT: No need any process for logout.',
+            })
+        }
     }
 }
